#204 refactor: 수정 - feed API 폴더 구주 및 참조 경로 수정
