--- front/src/api/feed/get-feed-detail.ts
@@ -1,6 +1,0 @@
-import { axios } from "@/lib/axios";
-import { FeedDetail } from "./types";
-
-export async function getFeedDetail(feedId: number): Promise<FeedDetail> {
-  return await axios.get(`/post/detail?postId=${feedId}`);
-}
