+++ front/src/api/feed/index.ts
@@ -1,7 +1,6 @@
-export { getFeedDetail } from "./get-feed-detail";
 export {
   useFeedList,
   usePrefetchFeedList,
   DEFAULT_POST_TYPE as DEFAULT_POST_TYPE,
   DEFAULT_SORT as DEFAULT_SORT,
-} from "./get-feed-list";
+} from "./hooks/get-feed-list";
